Convert db properties to javafx Rename to libraryproperties change separator width to make it visible
Remove small font setting on save order config
