+++ src/main/java/org/jabref/gui/Base.css
@@ -818,7 +818,7 @@
 
 .separator:horizontal .line {
     -fx-border-color: -jr-separator;
-    -fx-border-width: 0;
+    -fx-border-width: 0.5;
     -fx-border-insets: 1 0 0 0;
 }
